backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0